WiP on tests to match refactor [skip ci]
